% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shinyui.R
\name{addPlugin}
\alias{addPlugin}
\title{Add a plugin to the experiment viewer}
\usage{
addPlugin(id, ..., pluginList = default.plugins)
}
\arguments{
\item{id}{String to use for element id.}

\item{name}{A character name to use for the associated dashboard tab.}

\item{ui}{A function to call with the current experiment that returns a shiny
UI for the plugin tab.  The function should take the experiment object as
the only argument.}

\item{server}{A function that takes an experiment object as the only argument
and returns a shiny server function to handle the plugin.  The returned
shiny server function should take input, output and session arguments (in
that order).}

\item{icon}{Icon to use in the sidebar menu.}

\item{plugins}{A plugin set.  Defaults to the package wide plugin set.}
}
\description{
Add a plugin to the experiment viewer
}
\details{
See factors.R, doublingtime.R and apoindex.R in this package for
  example plugin implementations.
}

