% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/experiment.R
\name{appendData}
\alias{appendData}
\title{Load data from multiple source files.}
\usage{
appendData(paths, loader, design)
}
\arguments{
\item{paths}{Source files to load.}

\item{loader}{Function to use to load the source file.}

\item{design}{Design object, usually created with \code{\link{loadDesign}}.}
}
\value{
A validated data.frame with a 'time', 'well', and 'value' column.
}
\description{
This function calls \code{\link{loadData}} for multiple input files, appends
the results, and updates time stamps as direct in the design.
}
\details{
Usually there is no need to call this function directly; use
\code{\link{loadExperiment}} instead.

Several vignettes are included with this package to fully document
  the options for input files.  Run \code{vignette('source-files')} for a
  description of supported input file formats and loader functions.
}

