% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\docType{data}
\name{default.plugins}
\alias{default.plugins}
\title{Default viewer app plugins}
\format{An object of class \code{list} of length 6.}
\usage{
default.plugins
}
\description{
Default viewer app plugins
}
\keyword{datasets}

