% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/doublingtime.R
\docType{data}
\name{dt.filters}
\alias{dt.filters}
\title{Timepoint filter plugins}
\format{\preformatted{List of 2
 $ add:function (id, ...)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 6 15 8 15 15 15 6 8
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x2a03038> 
 $ get:function (id = names(plugins))  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 9 15 9 59 15 59 9 9
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x2a03038> 
}}
\usage{
dt.filters
}
\description{
Timepoint filter plugins
}
\keyword{datasets}

