% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublingtime.R
\docType{data}
\name{dt.timeFormats}
\alias{dt.timeFormats}
\title{Time conversion plugins}
\format{An object of class \code{list} of length 3.}
\usage{
dt.timeFormats
}
\description{
Time conversion plugins
}
\keyword{datasets}

