% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shinyui.R
\name{listPlugins}
\alias{listPlugins}
\title{List installed experiment viewer plugins}
\usage{
listPlugins(plugins = default.plugins)
}
\arguments{
\item{plugins}{A plugin set.  Defaults to the package wide set.}
}
\value{
A list of plugins
}
\description{
List installed experiment viewer plugins
}

