% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/experiment.R
\name{loadData}
\alias{loadData}
\title{Load, validate, and reformat (if necessary) source data.}
\usage{
loadData(path, parser, design)
}
\arguments{
\item{path}{Source file to load.}

\item{parser}{Function to use to load the source file.}

\item{design}{Design object, usually created with \code{\link{loadDesign}}.}
}
\value{
A validated data.frame with a 'time', 'well', and 'value' column.
}
\description{
This function loads data using the given reader and then validates that it
conforms with the expectations of the experimental design and reformats it as
needed.
}
\details{
Usually there is no need to call this function directly; use
\code{\link{loadExperiment}} instead.

Several vignettes are included with this package to fully document
  the options for input files.  Run \code{vignette('source-files')} for a
  description of supported input file formats and parser functions.
}

