% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/experiment.R
\name{loadDesign}
\alias{loadDesign}
\title{Load a design from a file or R list}
\usage{
loadDesign(path, design = yaml::yaml.load_file(path),
  default = writeDesign(file = NULL))
}
\arguments{
\item{path}{Path to a yaml design file.  Only used if \code{design} is not
given directly.}

\item{design}{A preliminary design data structure (See the
\link[=doc/design.html]{Design file} vignette for complete documentation of
the structure of a design).}

\item{default}{A design object to use as a source of default values to fill
in missing keys in \code{design}.}
}
\description{
This function should always be used to load design information, whether from
a yaml file or R list structure, because it performs various validation and
consistency checks.  It also injects reasonable defaults for missing
parameters.
}

