% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/experiment.R
\name{mergeData}
\alias{mergeData}
\title{Load and merge data for all channels.}
\usage{
mergeData(path, design)
}
\arguments{
\item{path}{A path to a directory where source files can be found.}

\item{design}{Design object, usually created with \code{\link{loadDesign}}.}
}
\value{
A data.frame containing a 'well', 'time' column and value columns for
  each channel.
}
\description{
This function loads and merges source data for all channels in an experiment.
}
\details{
Usually there is no need to call this function directly; use
\code{\link{loadExperiment}} instead.

Several vignettes are included with this package to fully document
  the options for input files.  Run \code{vignette('source-files')} for a
  description of supported input file formats and loader functions.
}

