% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/experiment.R
\name{newDesign}
\alias{newDesign}
\title{Write a new design object out to a yaml file.}
\usage{
newDesign(design = list(loader = "read.table", platform = "96", wells =
  "A1 -> H12", channels = list(channel1 = "data1.txt", channel2 = "data2.txt"),
  factors = list(factor1 = list(`A1->H6` = "A", `A7->H12` = "B"), factor2 =
  list(`A1->D12` = "C", `E1->H12` = "D"))), file = "design.yaml")
}
\arguments{
\item{design}{A optional design object.  Defaults to a sample experimental
design.}

\item{file}{An optional file path to use for creating a yaml text file.
  containing \code{design}.  Defaults to 'design.yaml' in the current working
  directory.}
}
\value{
A character vector containing yaml text.
}
\description{
This function serializes a design object out to a yaml text file or character
vector.  It is useful for writing a design object out to a file for future
use or for creating a template design file.
}

