% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateKinetics.R
\docType{package}
\name{plateKinetics}
\alias{plateKinetics}
\alias{plateKinetics-package}
\title{Analysis of kinetic reads from multiwell plates}
\description{
This package provides a framework for loading and analyzing data collected in
multiwell plate format over time.
}
\details{
A number of vignettes are provided to get you started.  Run 
\code{vignette('getting-started')} to see a simple getting started guide. See
\code{vignette('design-files')} for a description of the design file format. 
See \code{vignette('source-files')} for a description of supported input file
formats and loader functions.

If you'd rather not RTFM, jump to \code{\link{loadExperiment}} which load an 
experiment object which can be passed to downstream visualization and 
analysis functions.
}

