% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/experiment.R
\docType{data}
\name{platforms}
\alias{platforms}
\title{Platforms}
\format{\preformatted{List of 6
 $ 6   : chr [1:2, 1:3] "A1" "B1" "A2" "B2" ...
 $ 24  : chr [1:4, 1:6] "A1" "B1" "C1" "D1" ...
 $ 96  : chr [1:8, 1:12] "A1" "B1" "C1" "D1" ...
 $ 384 : chr [1:16, 1:24] "A1" "B1" "C1" "D1" ...
 $ 1536: chr [1:32, 1:48] "A1" "B1" "C1" "D1" ...
 $ 6144: chr [1:64, 1:96] "A1" "B1" "C1" "D1" ...
}}
\usage{
platforms
}
\description{
Platforms
}
\keyword{datasets}

