% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublingtime.R
\name{plotDoublingTimeFacet}
\alias{plotDoublingTimeFacet}
\title{Plot a value across time for all wells in the experiment.}
\usage{
plotDoublingTimeFacet(experiment, value, included)
}
\arguments{
\item{experiment}{An experiment object returned from 
\code{\link{doublingTime}}}

\item{value}{The name of a value on experiment$map that holds optical density
values.}

\item{included}{The name of the boolean column on experiment$map that records
which time points are used for the calculation.}
}
\value{
A ggplot2 object.
}
\description{
Plot a value across time for all wells in the experiment.
}

