% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/factors.R
\name{plotFactor}
\alias{plotFactor}
\title{Create a raster plot summarizing the levels/values for a factor in an
experiment.}
\usage{
plotFactor(experiment, factorName)
}
\arguments{
\item{experiment}{An experiment object, usually created by
\code{\link{loadExperiment}}.}

\item{factorName}{A character name giving the factor to be visulaized.}
}
\value{
A ggplot2 object.
}
\description{
Create a raster plot summarizing the levels/values for a factor in an
experiment.
}

