% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/map-viewer.R
\name{plotMapFacet}
\alias{plotMapFacet}
\title{Plot a value across time for all wells in the experiment.}
\usage{
plotMapFacet(experiment, value)
}
\arguments{
\item{experiment}{An experiment object, usually created by
\code{\link{loadExperiment}}.}

\item{value}{The name of a value on experiment$map to plot.}
}
\value{
A ggplot2 object.
}
\description{
Plot a value across time for all wells in the experiment.
}

