% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{template}
\alias{template}
\title{Copy files from an example project}
\usage{
template(template = "default", dest = getwd())
}
\arguments{
\item{template}{The name of the example to use.  Current options are
'default' for a project with space delimited input files or 'magellan' for
a project where data are read with the \code{\link{read.magellan}}
function.}

\item{dest}{The destination path.  Defaults to current working directory.}
}
\description{
Calling this function will not overwrite existing files by the same name.
}

