% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shinyui.R
\name{viewExperiment}
\alias{viewExperiment}
\title{View an experiment in an interactive shiny app}
\usage{
viewExperiment(experiment, plugins = default.plugins)
}
\arguments{
\item{experiment}{An experiment object, usually created by
\code{\link{loadExperiment}}.}

\item{plugins}{A set of plugins.  Defaults to the package wide set.}
}
\value{
An experiment object, with any updates made using the interactive
  app.
}
\description{
The experiment viewer is an interactive dashboard style shiny app that is
dynamically created when this function is called.
}

