% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublingtime.R
\docType{data}
\name{dt.filters}
\alias{dt.filters}
\title{Timepoint filter plugins}
\format{An object of class \code{list} of length 3.}
\usage{
dt.filters
}
\description{
Timepoint filter plugins
}
\keyword{datasets}

