% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/experiment.R
\name{loadExperiment}
\alias{loadExperiment}
\title{Load an experimental design and data files from a directory}
\usage{
loadExperiment(projectPath = getwd(),
  design = loadDesign(pickFile(projectPath, "design.yaml")))
}
\description{
Load an experimental design and data files from a directory
}

