% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\docType{data}
\name{platforms}
\alias{platforms}
\title{Standard multiwell plate platforms}
\format{An object of class \code{list} of length 7.}
\usage{
platforms
}
\description{
Built-in data structure containing default platforms, which are character
matrixes containing well lables for standard multiwell plate formats.
}
\details{
Designs files can directly reference names on this list for the
\code{platform} parameter (for example:  '96' or '384').  Run
\code{vignette('design-files')} for details.
}
\seealso{
\code{\link{platformLabels}} for creating additional platform 
  definitions.
}
\keyword{datasets}

