% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parsers.R
\name{read.magellan}
\alias{read.magellan}
\title{Load data from Megellan ASCI export files}
\usage{
read.magellan(file)
}
\arguments{
\item{file}{Path to the file to read.}
}
\value{
A data.frame in which the first column contains time stamps and
  subsequent columns readings from wells, row-wise across the plate (eg
  A1..A12 then B1..B12, etc.).
}
\description{
To use this read function to read data from files generated using Tecan's
Magellen suite, use the following settings for ASCI export files in your
method:
}
\details{
\describe{
  \item{Direction}{Horizontal}
  \item{Result}{Table (well data in columns)}
  \item{Add kinetic time stamps}{(checked)}
}
}

